Es remodelen les avaluacions per a que es facin a partir de les comandes. Es milloren les funcions de cercar establiments amb filtres i a més es retornen les estadistiques. S'ha implantat el autologin i la funció necessaria per a afegir més d'un element al rebost

+++ serveis/comandes.js
@@ -1,7 +1,12 @@
 const Comandes = require('../models/comandes.js');
+const mongoose = require('mongoose');
+const ObjectId = mongoose.Types.ObjectId;
+const OfertaServices = require('../serveis/ofertes');
 
 async function getAllComandesByUser(userId) {
-  let comandes = await Comandes.find({ userId });
+  let comandes = await Comandes.find({ userId: new ObjectId(userId) }).populate(
+    { path: 'establimentId' }
+  );
   if (!comandes) throw '404';
   return comandes;
 }
@@ -10,18 +15,54 @@ async function getAllComandesByOferta(ofertaId) {
   if (!comandes) throw '404';
   return comandes;
 }
-async function createComanda(comandaInfo) {
-  let comanda = new Comandes(comandaInfo);
-  return await comanda.save();
+async function getComanda(comandaId, userId) {
+  let comanda = await Comandes.find({
+    _id: comandaId,
+    userId: new ObjectId(userId),
+  });
+  if (!comanda) throw '404';
+  return comanda;
 }
-async function updateComanda() {}
+async function createComanda(comandaInfo, userId) {
+  console.log(userId);
+  if (
+    OfertaServices.quantitatOferta(
+      comandaInfo.establimentId,
+      comandaInfo.ofertaId,
+      comandaInfo.quantitat
+    )
+  ) {
+    let comanda = new Comandes({
+      userId: new ObjectId(userId),
+      data: Date.now(),
+      ...comandaInfo,
+    });
+    OfertaServices.restarQuantitatOferta(
+      comandaInfo.establimentId,
+      comandaInfo.ofertaId,
+      comandaInfo.quantitat
+    );
+    return await comanda.save();
+  }
+  throw '400';
+}
+async function updateComanda(comandaInfo, comandaId, userId) {
+  let comanda = await Comandes.findOneAndUpdate(
+    { _id: comandaId, userId: new ObjectId(userId) },
+    {
+      $set: { $: comandaInfo },
+    }
+  );
+  if (!comanda) throw '404';
+  return comanda;
+}
 async function deleteComanda(comandaId) {
   let deleted = await Comandes.findOneAndDelete({ _id: comandaId });
   if (!deleted) throw '404';
   return deleted;
 }
 async function deleteComandesByUser(userId) {
-  let deleted = await Comandes.deleteMany({ userId });
+  let deleted = await Comandes.deleteMany({ userId: new ObjectId(userId) });
   if (!deleted) throw '404';
   return deleted;
 }
@@ -34,6 +75,7 @@ async function deleteComandesByOferta(ofertaId) {
 module.exports = {
   getAllComandesByUser,
   getAllComandesByOferta,
+  getComanda,
   createComanda,
   updateComanda,
   deleteComanda,
